requirements.txt
pillow
django-ckeditor
Django==5.0.6